requirements.txt
colorama==0.4.6
docx2pdf==0.1.8
lxml==6.0.2
PyMuPDF==1.26.6
python-docx==1.2.0
pywin32==311
tqdm==4.67.1
typing_extensions==4.15.0
